move the command to src/commands and add exports
