+++ packages/plugin-gate/src/cli.ts
@@ -1,3 +1,3 @@
-import {GateEvaluateCommand} from './evaluate'
+import {PluginCommand as GateEvaluateCommand} from './commands/evaluate'
 
 module.exports = [GateEvaluateCommand]
