+++ packages/plugin-gate/src/commands/evaluate.ts
@@ -1,6 +1,7 @@
-import type {APIHelper, EvaluationResponse, EvaluationResponsePayload, Payload, PayloadOptions} from './interfaces'
+import type {APIHelper, EvaluationResponse, EvaluationResponsePayload, Payload, PayloadOptions} from '../interfaces'
 import type {AxiosResponse} from 'axios'
 
+import {GateEvaluateCommand} from '@datadog/datadog-ci-base/commands/gate/evaluate-command'
 import {FIPS_ENV_VAR, FIPS_IGNORE_ERROR_ENV_VAR} from '@datadog/datadog-ci-base/constants'
 import {getCISpanTags} from '@datadog/datadog-ci-base/helpers/ci'
 import {toBoolean} from '@datadog/datadog-ci-base/helpers/env'
@@ -8,84 +9,21 @@ import {enableFips} from '@datadog/datadog-ci-base/helpers/fips'
 import {getGitMetadata} from '@datadog/datadog-ci-base/helpers/git/format-git-span-data'
 import {SpanTags} from '@datadog/datadog-ci-base/helpers/interfaces'
 import {Logger, LogLevel} from '@datadog/datadog-ci-base/helpers/logger'
-import {executePluginCommand} from '@datadog/datadog-ci-base/helpers/plugin'
 import {retryRequest} from '@datadog/datadog-ci-base/helpers/retry'
 import {GIT_HEAD_SHA, GIT_PULL_REQUEST_BASE_BRANCH, parseTags} from '@datadog/datadog-ci-base/helpers/tags'
 import {getUserGitSpanTags} from '@datadog/datadog-ci-base/helpers/user-provided-git'
-import * as validation from '@datadog/datadog-ci-base/helpers/validation'
 import chalk from 'chalk'
-import {Command, Option} from 'clipanion'
 import {v4 as uuidv4} from 'uuid'
 
-import {apiConstructor} from './api'
+import {apiConstructor} from '../api'
 import {
   renderEvaluationResponse,
   renderGateEvaluationInput,
   renderGateEvaluationError,
   renderEvaluationRetry,
   renderWaiting,
-} from './renderer'
-import {getBaseIntakeUrl, is4xxError, is5xxError, isTimeout, parseScope} from './utils'
-
-export class GateEvaluateCommand extends Command {
-  public static paths = [['gate', 'evaluate']]
-
-  public static usage = Command.Usage({
-    category: 'CI Visibility',
-    description: 'Evaluate Quality Gates rules in Datadog.',
-    details: `
-      This command will evaluate the matching quality gate rules in Datadog.\n
-      See README for details.
-    `,
-    examples: [
-      ['Evaluate matching quality gate rules in Datadog', 'datadog-ci gate evaluate'],
-      [
-        'Evaluate matching quality gate rules in Datadog, failing if no rules were found',
-        'datadog-ci gate evaluate --fail-on-empty',
-      ],
-      [
-        'Evaluate matching quality gate rules in Datadog, failing if Datadog is not available',
-        'datadog-ci gate evaluate --fail-if-unavailable',
-      ],
-      [
-        'Evaluate matching quality gate rules in Datadog and add extra scope',
-        'datadog-ci gate evaluate --scope team:backend',
-      ],
-      [
-        'Evaluate matching quality gate rules in Datadog and add extra tags',
-        'datadog-ci gate evaluate --tags team:frontend',
-      ],
-      [
-        'Evaluate matching quality gate rules in Datadog from the datadoghq.eu site',
-        'DD_SITE=datadoghq.eu datadog-ci gate evaluate',
-      ],
-      [
-        'Evaluate matching quality gate rules in Datadog with a timeout of 120 seconds',
-        'datadog-ci gate evaluate --timeout 120',
-      ],
-      ['Evaluate matching quality gate rules in Datadog without waiting', 'datadog-ci gate evaluate --no-wait'],
-    ],
-  })
-
-  private defaultTimeout = 600 // 10 min
-
-  protected dryRun = Option.Boolean('--dry-run', false)
-  protected failOnEmpty = Option.Boolean('--fail-on-empty', false)
-  protected failIfUnavailable = Option.Boolean('--fail-if-unavailable', false)
-  protected noWait = Option.Boolean('--no-wait', false)
-  protected timeoutInSeconds = Option.String('--timeout', String(this.defaultTimeout), {
-    validator: validation.isInteger(),
-  })
-  protected userScope = Option.Array('--scope')
-  protected tags = Option.Array('--tags')
-
-  protected fips = Option.Boolean('--fips', false)
-  protected fipsIgnoreError = Option.Boolean('--fips-ignore-error', false)
-
-  public async execute() {
-    return executePluginCommand(this)
-  }
-}
+} from '../renderer'
+import {getBaseIntakeUrl, is4xxError, is5xxError, isTimeout, parseScope} from '../utils'
 
 export class PluginCommand extends GateEvaluateCommand {
   private initialRetryMs = 1000
@@ -150,7 +88,7 @@ export class PluginCommand extends GateEvaluateCommand {
     return apiConstructor(getBaseIntakeUrl(), this.config.apiKey, this.config.appKey)
   }
 
-  protected async getSpanTags(): Promise<SpanTags> {
+  private async getSpanTags(): Promise<SpanTags> {
     const ciSpanTags = getCISpanTags()
     const gitSpanTags = await getGitMetadata()
     const userGitSpanTags = getUserGitSpanTags()
@@ -159,8 +97,8 @@ export class PluginCommand extends GateEvaluateCommand {
     const cliTags = this.tags ? parseTags(this.tags) : {}
 
     return {
-      ...gitSpanTags,
       ...ciSpanTags,
+      ...gitSpanTags,
       ...userGitSpanTags,
       ...cliTags,
       ...envVarTags,
@@ -168,7 +106,7 @@ export class PluginCommand extends GateEvaluateCommand {
 
   }
 
-  protected async evaluateRules(api: APIHelper, evaluateRequest: Payload): Promise<number> {
+  private async evaluateRules(api: APIHelper, evaluateRequest: Payload): Promise<number> {
     this.context.stdout.write(renderGateEvaluationInput(evaluateRequest))
 
     /**
@@ -185,8 +123,6 @@ export class PluginCommand extends GateEvaluateCommand {
         }
       },
       retries: this.maxRetries,
-      maxTimeout: 0,
-      minTimeout: 0,
     })
       .then((response) => {
         return this.handleEvaluationSuccess(response.data.data.attributes)
@@ -211,7 +147,8 @@ export class PluginCommand extends GateEvaluateCommand {
     bail?: (e: Error) => void
   ): Promise<AxiosResponse<EvaluationResponsePayload>> {
     const timePassed = new Date().getTime() - evaluateRequest.startTimeMs
-    const remainingWait = this.timeoutInSeconds * 1000 - timePassed
+    const timeoutInSeconds = parseInt(this.timeoutInSeconds, 10)
+    const remainingWait = timeoutInSeconds * 1000 - timePassed
 
     return new Promise((resolve, reject) => {
       const request = {...evaluateRequest, options: {...evaluateRequest.options}}
